fix(wasm): reject a memory larger than a slice instead of panicking
make panics rather than returning an error past math.MaxInt, so a module
declaring a memory bigger than that took the host down with
"makeslice: len out of range" -- a two-gibibyte memory did it on any
platform whose int is 32 bits wide (GOARCH=386, arm, wasm), and nothing
about the module has to be malformed to get there.

MaxAllocatablePages is the ceiling a memory can actually reach on the
running platform. On a 64-bit one it is the specification's own 65536
pages and nothing changes at all; on a 32-bit one it caps a memory just
under two gibibytes. The memory sizer clamps with it so the maximum it
fills in for an unbounded memory is one that can be honoured, and
Memory.Validate clamps with it so a declared minimum over the ceiling is
turned away before anything is allocated. Providing less memory than a
module asks for is something the specification lets a host do.

The two data-segment bounds checks had the int problem too: int(offset)
truncates the offset expression, and adding len(Init) to an offset near
MaxInt32 overflows to a negative ceiling that passes the check and
panics the copy. Both compare in uint64 now, as applyElements does.

Tests that named 65536 in an error or expected a decoded maximum now
derive it, so they assert the same behaviour at either width.

+++ internal/component/instance/stream_test.go
@@ -291,7 +291,7 @@ func TestStreamCopy_BlockedSentinelAsync(t *testing.T) {
 	asyncRead := streamCopyHostFunc(in, sideReadable, eventStreamRead, nil, 0, 0, true, true, nil, nil)
 	out := callBuiltin(asyncRead, uint64(r), 0, 1)
 	if uint32(out[0]) != blockedSentinel {
-		t.Fatalf("got %#x, want BLOCKED (%#x)", out[0], blockedSentinel)
+		t.Fatalf("got %#x, want BLOCKED (%#x)", out[0], uint32(blockedSentinel))
 	}
 }
